getPhyloNames_noCache

Internal helper function for getPhyloNames; looks for either a scientific or common name matching whichever you supply, based on nameType. Does the work of interfacing with the sci2comm and comm2sci functions

getPhyloNames_noCache(speciesNames, nameType, quiet = T)

Arguments

speciesNames
a name or vector of common or scientific names of organisms of interest (in quotes)

nameType
what type of name are you supplying? Either "sci" or "common"

quiet
suppress verbose feedback from the taxize package? Default=T

Details

You generally won't want or need to use this function, since getPhyloNames invokes it and will cache the results to run more efficiently. But this is basically a convenience wrapper that can suppress some of the messaging from taxize and also search both NCBI and EOL databases for taxonomic matches.