theme_galactic

A ggplot2 theme for Galactic Polymath styling. Sensible defaults for plots intended for presentations and worksheets. (Large text, thick grid lines, etc.)

theme_galactic(
  grid.thickness.maj = 0.7,
  grid.thickness.min = 0.4,
  grid.col = "#C3C3C3",
  border.thickness = 1,
  border.col = "#6D6D6D",
  font = "Montserrat",
  regular.wt = 400,
  bold.wt = 700,
  font.cex = 1,
  font.face = 1,
  axis.lab.col = "#363636",
  axis.text.col = "#6D6D6D",
  axis.tick.length = 10,
  plot.margin = ggplot2::margin(t = 20, r = 20, b = 10, l = 20)
)

Arguments

grid.thickness.maj
How heavy do you want grid lines to be? (in case printer makes things lighter); default=.8

grid.thickness.min
How heavy do you want grid lines to be? (in case printer makes things lighter); default=.6

grid.col
What color do you want the grid to be? Default: same as font (#363636)

border.thickness
How heavy do you want the plot border to be?

border.col
Color of plot border. Default: same as font (#363636)

font
Google font to use, "Montserrat" by default; see options with sysfonts::font_families_google()

regular.wt
font weight for regular font style

bold.wt
font weight for bold text

font.cex
a simple multiplier for scaling all text

font.face
style of axis label and title fonts; 1=plain, 2= bold, 3=italic, 4=bold+italic; Provide 1 value for all or 3 values for title, x-axis label, y-axis label (in that order); default= 1 (plain)

axis.lab.col
color of axis labels (and title)

axis.text.col
color of axis text (numbers, dates, etc)

axis.tick.length
length of axis ticks (in pt units)

plot.margin
easy access to ggplot margins